+++ src/dx/dx_utils.rs
@@ -1,0 +1,16 @@
+use std::collections::HashMap;
+
+
+pub fn default_str() -> String {
+    "".to_string()
+}
+pub fn default_map() -> HashMap<String, String> {
+    HashMap::new()
+}
+pub fn default_bool() -> bool {
+    false
+}
+#[allow(dead_code)]
+pub fn default_vec() -> Vec<String> {
+    Vec::new()
+}