+++ src/dx/mod.rs
@@ -1,0 +1,11 @@
+// Modules de gestion du dictionnaire
+
+
+mod dx_portail;
+pub use self::dx_portail::Portail;
+
+
+mod dx_application;
+
+mod dx_utils;
+